+++ Runtime/API/Glyph.cs
@@ -44,13 +44,16 @@ namespace LMirman.RewiredGlyphs
 		/// https://guavaman.com/projects/rewired/docs/HowTos.html#display-glyph-for-action
 		/// </remarks>
 		public int InputID => inputID;
+
 		/// <summary>
 		/// Deprecated description of the input that formerly represent an absolute description of the Glyph.
 		/// Superseded by <see cref="GetDescription"/> and <see cref="FullDescription"/>.
 		/// </summary>
 		/// <seealso cref="GetDescription"/><seealso cref="FullDescription"/>.
-		[Obsolete("Description is obsolete since it does not represent a particular axis. Use GetDescription(AxisRange) to get a dynamic description or FullDescription to explicitly get the AxisRange.Full description.")]
+		[Obsolete("Description is obsolete since it does not represent a particular axis. " +
+		          "Use GetDescription(AxisRange) to get a dynamic description or FullDescription to explicitly get the AxisRange.Full description.")]
 		public string Description => description;
+
 		/// <summary>
 		/// The description of this input for the device for the <see cref="AxisRange.Full"/> axis.
 		/// </summary>
@@ -59,6 +62,7 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		/// <seealso cref="GetDescription"/>
 		public string FullDescription => description;
+
 		/// <summary>
 		/// The description of this input for the device for the <see cref="AxisRange.Positive"/> axis.
 		/// </summary>
@@ -68,6 +72,7 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		/// <seealso cref="GetDescription"/>
 		public string PositiveDescription => positiveDescription;
+
 		/// <summary>
 		/// The description of this input for the device for the <see cref="AxisRange.Negative"/> axis.
 		/// </summary>
@@ -77,6 +82,7 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		/// <seealso cref="GetDescription"/>
 		public string NegativeDescription => negativeDescription;
+
 		/// <summary>
 		/// The sprite representing this input for the <see cref="AxisRange.Full"/> axis.
 		/// </summary>
@@ -85,6 +91,7 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		/// <seealso cref="GetSprite"/>
 		public Sprite FullSprite => sprite;
+
 		/// <summary>
 		/// The sprite representing this input for the <see cref="AxisRange.Positive"/> axis.
 		/// </summary>
@@ -95,6 +102,7 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		/// <seealso cref="GetSprite"/>
 		public Sprite PositiveSprite => positiveSprite;
+
 		/// <summary>
 		/// The sprite representing this input for the <see cref="AxisRange.Negative"/> axis.
 		/// </summary>
@@ -105,6 +113,7 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		/// <seealso cref="GetSprite"/>
 		public Sprite NegativeSprite => negativeSprite;
+
 		/// <summary>
 		/// The name of the text mesh sprite sheet that contains the sprites for this glyph.
 		/// </summary>
@@ -116,17 +125,20 @@ namespace LMirman.RewiredGlyphs
 			get => textMeshSpriteSheetName;
 			set => textMeshSpriteSheetName = value;
 		}
+
 		/// <summary>
 		/// When true this Glyph should show as text instead of its sprite.
 		/// </summary>
 		public bool PreferDescription { get; set; }
+
 		/// <summary>
 		/// What this Glyph is intended to represent at runtime.
 		/// </summary>
 		/// <seealso cref="Type"/>
 		public Type GlyphType { get; internal set; }
+
 		/// <summary>
-		/// The <see cref="Rewired.ControllerType"/> that this Glyph intends to be represent.
+		/// The <see cref="Rewired.ControllerType"/> that this Glyph intends to represent.
 		/// </summary>
 		/// <remarks>
 		/// <b>Important:</b> Depending on the value of <see cref="GlyphType"/> this value may be null if it doesn't represent a specific device.<br/><br/>
@@ -134,10 +146,12 @@ namespace LMirman.RewiredGlyphs
 		/// </remarks>
 		[CanBeNull]
 		public ControllerType? ControllerType { get; internal set; }
+
 		/// <summary>
 		/// True when <see cref="GlyphType"/> is <see cref="Type.Input"/>.
 		/// </summary>
 		public bool IsInputGlyph => GlyphType == Type.Input;
+
 		/// <summary>
 		/// True when <see cref="GlyphType"/> is <i>not</i> <see cref="Type.Undefined"/> or <see cref="Type.Input"/>.
 		/// </summary>
@@ -186,6 +200,50 @@ namespace LMirman.RewiredGlyphs
 			}
 		}
 
+		public Glyph(int inputID, string description, ControllerType? controllerType, Sprite sprite, Sprite positiveSprite, Sprite negativeSprite, Type type = Type.Input)
+		{
+			this.inputID = inputID;
+			this.description = description;
+			this.sprite = sprite;
+			this.positiveSprite = positiveSprite;
+			this.negativeSprite = negativeSprite;
+			PreferDescription = false;
+			GlyphType = type;
+			ControllerType = controllerType;
+		}
+
+		public Glyph(int inputID, string description, ControllerType? controllerType, Sprite sprite, Type type = Type.Input)
+		{
+			this.inputID = inputID;
+			this.description = description;
+			this.sprite = sprite;
+			positiveSprite = null;
+			negativeSprite = null;
+			PreferDescription = sprite != null;
+			GlyphType = type;
+			ControllerType = controllerType;
+		}
+
+		/// <summary>
+		/// Create a fallback glyph that only contains a description.
+		/// </summary>
+		/// <remarks>The main use case is if there is no glyph found but there is <i>some</i> information about the input that can allow it to be shown in text.</remarks>
+		public Glyph(string description, ControllerType? controllerType, Sprite sprite = null, Type type = Type.Input)
+		{
+			inputID = -1;
+			this.description = description;
+			this.sprite = sprite;
+			positiveSprite = null;
+			negativeSprite = null;
+			PreferDescription = true;
+			GlyphType = type;
+			ControllerType = controllerType;
+		}
+
+		#region Obsolete
+		/// <inheritdoc cref="Glyph(int, string, Rewired.ControllerType?, Sprite, Sprite, Sprite, Type)"/>
+		[Obsolete("This constructor does not specify ControllerType which may cause issues with other systems. " +
+		          "Use Glyph(int, string, Rewired.ControllerType, Sprite, Sprite, Sprite, Type) instead.")]
 		public Glyph(int inputID, string description, Sprite sprite, Sprite positiveSprite, Sprite negativeSprite, Type type = Type.Input)
 		{
 			this.inputID = inputID;
@@ -197,6 +255,9 @@ namespace LMirman.RewiredGlyphs
 			GlyphType = type;
 		}
 
+		/// <inheritdoc cref="Glyph(int, string, Rewired.ControllerType?, Sprite, Type)"/>
+		[Obsolete("This constructor does not specify ControllerType which may cause issues with other systems. " +
+		          "Glyph(int, string, Rewired.ControllerType, Sprite, Type) instead.")]
 		public Glyph(int inputID, string description, Sprite sprite, Type type = Type.Input)
 		{
 			this.inputID = inputID;
@@ -208,10 +269,9 @@ namespace LMirman.RewiredGlyphs
 			GlyphType = type;
 		}
 
-		/// <summary>
-		/// Create a fallback glyph that only contains a description.
-		/// </summary>
-		/// <remarks>The main use case is if there is no glyph found but there is <i>some</i> information about the input that can allow it to be shown in text.</remarks>
+		/// <inheritdoc cref="Glyph(string, Rewired.ControllerType?, Sprite, Type)"/>
+		[Obsolete("This constructor does not specify ControllerType which may cause issues with other systems. " +
+		          "Use Glyph(string, ControllerType) instead.")]
 		public Glyph(string description, Sprite sprite = null, Type type = Type.Input)
 		{
 			inputID = -1;
@@ -222,6 +282,7 @@ namespace LMirman.RewiredGlyphs
 			PreferDescription = true;
 			GlyphType = type;
 		}
+		#endregion
 
 		[PublicAPI]
 		public enum Type